Add a --dry-run option
Add a --dry-run option, similar in nature to common Linux shell
scripts.  If this flag is specified, polly will create the generate
command, print it to the log and stdout, and then quit.
This allows one to use familiar polly one-liners to create
equivalent CMake generate commands.  This can be convenient
for a number of reasons, such as troubleshooting builds or creating
toolchain configurations for CMake friendly IDE's.  This provides
a convenient alternative to starting a build, killing it (ctrl-c)
and then grepping the log.